+++ modules/pytket-quantinuum/docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 * Add wasm support
 * Add support for `OpType.CopyBits` and `OpType.ClassicalExpBox` in `QuantinuumBackend`
 * Updated pytket version requirement to 1.3.0rc1
+* Add optional argument `group` to `QuantinuumBackend`
 
 0.3.1 (May 2022)
 ----------------